# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: dab_team@T-Mobile.com
    name: DAB Team
  description: Services related APIs
  title: Services
  version: 1.0.0
  x-type-of-consumer: EU
  x-breadth-of-consumption: ANY
  x-data-sensitivity: Restricted
servers:
- url: https://dab-plans-services:9000/v1
paths:
  /services:
    post:
      description: Get current and eligible services including data passes
      operationId: services
      parameters:
      - in: header
        name: Accept-Language
        schema:
          type: string
      - description: Bearer prefix JWT Token
        in: header
        name: Authorization
        required: true
        schema:
          pattern: .*$
          type: string
      - description: Taap compliant POPT Token.
        in: header
        name: X-Authorization
        required: true
        schema:
          pattern: .*$
          type: string
      - description: Id token.
        in: header
        name: X-Auth-Originator
        required: true
        schema:
          pattern: .*$
          type: string
      - in: header
        name: app-version
        schema:
          type: string
      - in: header
        name: device-oem
        schema:
          type: string
      - in: header
        name: device-model
        schema:
          type: string
      - in: header
        name: device-os
        schema:
          type: string
      - in: header
        name: device-os-version
        schema:
          type: string
      - in: header
        name: os-language
        schema:
          type: string
      - in: header
        name: timestamp
        schema:
          type: string
      - in: header
        name: channel-id
        schema:
          type: string
      - in: header
        name: origin-application-id
        schema:
          type: string
      - in: header
        name: application-id
        schema:
          type: string
      - in: header
        name: session-id
        schema:
          type: string
      - in: header
        name: interaction-id
        schema:
          type: string
      - in: header
        name: activity-id
        schema:
          type: string
      - in: header
        name: Content-Type
        required: true
        schema:
          pattern: .*$
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicesRequest'
        description: services
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesResponse'
          description: services
          headers:
            Cache-Control:
              description: Cache-control
              schema:
                pattern: .*$
                type: string
            Date:
              description: time stamp of response generation
              schema:
                pattern: .*$
                type: string
        "400":
          content: {}
          description: Bad request
        "401":
          content: {}
          description: Unauthorized
        "403":
          content: {}
          description: Client access denied
        "500":
          content: {}
          description: Internal server error
      security:
      - OAuth2: []
      summary: Get services
      tags:
      - services
      x-data-sensitivity: Restricted
      x-api-pattern: ExecuteFunction
      x-codegen-request-body-name: services
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: services
...
  /update-services:
    post:
      operationId: api to update plan/services/datapass
      parameters:
      - in: header
        name: Accept-Language
        schema:
          type: string
      - description: Bearer prefix JWT Token
        in: header
        name: Authorization
        required: true
        schema:
          pattern: .*$
          type: string
      - description: Taap compliant POPT Token.
        in: header
        name: X-Authorization
        required: true
        schema:
          pattern: .*$
          type: string
      - description: Id token.
        in: header
        name: X-Auth-Originator
        required: true
        schema:
          pattern: .*$
          type: string
      - in: header
        name: app-version
        schema:
          type: string
      - in: header
        name: device-oem
        schema:
          type: string
      - in: header
        name: device-model
        schema:
          type: string
      - in: header
        name: device-os
        schema:
          type: string
      - in: header
        name: device-os-version
        schema:
          type: string
      - in: header
        name: os-language
        schema:
          type: string
      - in: header
        name: timestamp
        schema:
          type: string
      - in: header
        name: channel-id
        schema:
          type: string
      - in: header
        name: origin-application-id
        schema:
          type: string
      - in: header
        name: application-id
        schema:
          type: string
      - in: header
        name: session-id
        schema:
          type: string
      - in: header
        name: interaction-id
        schema:
          type: string
      - in: header
        name: activity-id
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePlanRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePlanResponse'
          description: Ok
          headers:
            Cache-Control:
              description: Cache policy supported for the resource by the server
              schema:
                maxLength: 256
                minLength: 1
                pattern: "^[\\S]*$"
                type: string
            service-transaction-id:
              description: Request ID echoed back by server
              schema:
                maxLength: 256
                minLength: 1
                pattern: "^[\\S]*$"
                type: string
        "400":
          content:
            application/json:
              example:
                code: "400"
                userMessage: Bad Request
                systemMessage: Bad Request
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
        "401":
          content:
            application/json:
              example:
                code: "401"
                userMessage: Unauthorized
                systemMessage: Unauthorized
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
        "403":
          content:
            application/json:
              example:
                exception: "To read more details about this error, visit https://qwiki.internal.t-mobile.com/display/DIGEE/Publisher+-+Error+codes\
                  \ and look for error the specific error code."
                path: "ServletWebRequest: uri=/development/v1/events/DemoEvent;client=198.18.53.102"
                message: "To read more details about this error, visit https://qwiki.internal.t-mobile.com/display/DIGEE/Publisher+-+Error+codes\
                  \ and look for error the specific error code."
                error: FORBIDDEN
                timestamp: Thu Jun 27 22:35:14 UTC 2019
                status: "403"
              schema:
                $ref: '#/components/schemas/Error'
          description: ProducerId is not authorized to publish event in env or Event
            not associated with any Team name in DEEP.io Registry
        "404":
          content:
            application/json:
              example:
                code: "404"
                userMessage: Resource not found
                systemMessage: Resource not found
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Resource not found
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
        "405":
          content:
            application/json:
              example:
                code: "405"
                userMessage: Method Not Allowed
                systemMessage: Method Not Allowed
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Method Not Allowed
          headers:
            service-transaction-id:
              description: Request ID echoed back by server
              schema:
                maxLength: 256
                minLength: 1
                pattern: "^[\\S]*$"
                type: string
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
            Allow:
              description: list of supported methods for URI
              schema:
                maxLength: 256
                minLength: 1
                pattern: "^[\\S]*$"
                type: string
        "406":
          content:
            application/json:
              example:
                code: "406"
                userMessage: Mismatching data format
                systemMessage: Mismatching data format
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Mismatching data format
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
        "500":
          content:
            application/json:
              example:
                code: "500"
                userMessage: System Error
                systemMessage: System Error
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: System Error
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
        "503":
          content:
            application/json:
              example:
                code: "503"
                userMessage: Service unavailable
                systemMessage: Service unavailable
                detailLink: http://www.tmus.com
              schema:
                $ref: '#/components/schemas/Error'
          description: Service unavailable
          headers:
            Date:
              description: The Date general HTTP header contains the date and time
                at which the message was originated.
              schema:
                format: date-time
                type: string
      security:
      - OAuth2: []
      summary: Change plan and services
      tags:
      - services
      x-api-pattern: queryResource
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: services
...
components:
...
  schemas:
    Error:
      description: Error object
      example:
        userMessage: userMessage
        systemMessage: systemMessage
        code: code
      properties:
        code:
          description: code for error
          type: string
        systemMessage:
          description: system message for error
          type: string
        userMessage:
          description: user message
          type: string
      type: object
    ServicesRequest:
      description: Request to get services
      example:
        serviceSubTypeFilter: Handset Protection and Upgrade/Video Streaming - only
          for services and not subscriptions
        fetchSubscriptions: true
        msisdn: "1234567890"
        subCategoriesFilter:
        - Insurance
        - Extended Warranty
        - Mobile Security
        - Upgrade Program
        fetchPlanDataServices: true
        fetchDataPasses: true
      properties:
        msisdn:
          example: "1234567890"
          type: string
        fetchSubscriptions:
          example: true
          type: boolean
        fetchDataPasses:
          example: true
          type: boolean
        fetchPlanDataServices:
          example: true
          type: boolean
        serviceSubTypeFilter:
          example: Handset Protection and Upgrade/Video Streaming - only for services
            and not subscriptions
          type: string
        subCategoriesFilter:
          example:
          - Insurance
          - Extended Warranty
          - Mobile Security
          - Upgrade Program
          items:
            type: string
          type: array
      type: object
...
    ChangePlanRequest:
      description: request for change plan/service/datapass
      example:
        serviceChanges:
        - productLevel: true
          accountLevel: true
          productType: GSM
          changeDetails:
            sequenceNumber: 700064111
            subCategory: AppleTVplus
            soc: HDPASS
            partnerTermsAndConditionText: "By checking the box, you agree to activate\
              \ Netflix, a third-party service. T-Mobile will handle billing and payments\
              \ for this service; recurring charges will appear on your T-Mobile bill.\
              \ T-Mobile’s Terms and Conditions, including arbitration of disputes,\
              \ apply to this transaction. You will also need to agree to the <strong><a\
              \ class=\\\"link-secondary\\\" href=\\\"https://help.netflix.com/legal/termsofuse?locale=en-US\\\
              \" target=\\\"_blank\\\">Netflix Subscription Terms</a></strong> before\
              \ using Netflix."
            changeType: changeDataPass
            action: ADD
            serviceSubType: Video Streaming
            autoRenewalTermsAndConditionText: "By checking the box, you agree that\
              \ after your 4-month promotional trial concludes, you will automatically\
              \ be charged $9.99 per month on your T-Mobile bill. You may cancel at\
              \ any time during the trial."
            socName: Apple TV+ On Us
            effectiveDate: 2020-03-20T00:00:00.000Z
            expirationDate: 2020-03-20T00:00:00.000Z
        - productLevel: true
          accountLevel: true
          productType: GSM
          changeDetails:
            sequenceNumber: 700064111
            subCategory: AppleTVplus
            soc: HDPASS
            partnerTermsAndConditionText: "By checking the box, you agree to activate\
              \ Netflix, a third-party service. T-Mobile will handle billing and payments\
              \ for this service; recurring charges will appear on your T-Mobile bill.\
              \ T-Mobile’s Terms and Conditions, including arbitration of disputes,\
              \ apply to this transaction. You will also need to agree to the <strong><a\
              \ class=\\\"link-secondary\\\" href=\\\"https://help.netflix.com/legal/termsofuse?locale=en-US\\\
              \" target=\\\"_blank\\\">Netflix Subscription Terms</a></strong> before\
              \ using Netflix."
            changeType: changeDataPass
            action: ADD
            serviceSubType: Video Streaming
            autoRenewalTermsAndConditionText: "By checking the box, you agree that\
              \ after your 4-month promotional trial concludes, you will automatically\
              \ be charged $9.99 per month on your T-Mobile bill. You may cancel at\
              \ any time during the trial."
            socName: Apple TV+ On Us
            effectiveDate: 2020-03-20T00:00:00.000Z
            expirationDate: 2020-03-20T00:00:00.000Z
        msisdn: "1234567890"
      properties:
        msisdn:
          description: MSISDN in action
          example: "1234567890"
          nullable: true
          pattern: "^\\d{10}$"
          type: string
        serviceChanges:
          description: list of services to be changed
          items:
            $ref: '#/components/schemas/ServiceChange'
          type: array
      required:
      - msisdn
      - serviceChanges
      type: object
    ServiceChange:
      example:
        productLevel: true
        accountLevel: true
        productType: GSM
        changeDetails:
          sequenceNumber: 700064111
          subCategory: AppleTVplus
          soc: HDPASS
          partnerTermsAndConditionText: "By checking the box, you agree to activate\
            \ Netflix, a third-party service. T-Mobile will handle billing and payments\
            \ for this service; recurring charges will appear on your T-Mobile bill.\
            \ T-Mobile’s Terms and Conditions, including arbitration of disputes,\
            \ apply to this transaction. You will also need to agree to the <strong><a\
            \ class=\\\"link-secondary\\\" href=\\\"https://help.netflix.com/legal/termsofuse?locale=en-US\\\
            \" target=\\\"_blank\\\">Netflix Subscription Terms</a></strong> before\
            \ using Netflix."
          changeType: changeDataPass
          action: ADD
          serviceSubType: Video Streaming
          autoRenewalTermsAndConditionText: "By checking the box, you agree that after\
            \ your 4-month promotional trial concludes, you will automatically be\
            \ charged $9.99 per month on your T-Mobile bill. You may cancel at any\
            \ time during the trial."
          socName: Apple TV+ On Us
          effectiveDate: 2020-03-20T00:00:00.000Z
          expirationDate: 2020-03-20T00:00:00.000Z
      properties:
        accountLevel:
          description: Is change an account level or line level
          example: true
          type: boolean
        productLevel:
          description: Is change an product level or not
          example: true
          type: boolean
        productType:
          description: type of product
          example: GSM
          type: string
        changeDetails:
          $ref: '#/components/schemas/ChangeDetails'
      required:
      - changeDetails
      type: object
...
    ChangePlanResponse:
      example:
        statusDetails:
          status: SUCCESS
          statusCode: "100"
      properties:
        statusDetails:
          $ref: '#/components/schemas/StatusDetails'
      type: object
    StatusDetails:
      example:
        status: SUCCESS
        statusCode: "100"
      properties:
        status:
          description: This field represent status of the service(SUCCESS or ERROR).
          example: SUCCESS
          type: string
        statusCode:
          description: "This field represent status code(200 , 404 , 500)"
          example: "100"
          type: string
      type: object
...
  securitySchemes:
    OAuth2:
      description: Access via JWT
      in: header
      name: Authorization
      type: apiKey
...